Rename AWSLocalClusterDiscovery + Added VPC public and localhost discovery impl
